requirements.txt
asgiref==3.2.10
automium==0.2.6
automium-web==0.1.1
click==7.1.2
Django==2.2.3
django-filter==2.3.0
djangorestframework==3.11.0
Flask==1.1.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
legacy==0.1.6
MarkupSafe==1.1.1
PyMySQL==0.9.3
pytz==2020.1
sqlparse==0.3.1
Werkzeug==1.0.1
whitenoise==5.2.0
